undo most of the refactor in "many: configure store from state, reconfigure store at runtime (#3780)"
we plan to resolve the via (store) id much more dynamicall in the end, more like what we to with the model assertion
and the store in it

This reverts most of commit 76b70d18f21e76e634ef880250ef62d4d2e3c0de.
