+++ overlord/overlord_test.go
@@ -21,7 +21,6 @@ package overlord_test
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -39,6 +38,7 @@ import (
 	"github.com/snapcore/snapd/overlord/patch"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
+	"github.com/snapcore/snapd/store"
 	"github.com/snapcore/snapd/testutil"
 )
 
@@ -63,12 +63,6 @@ func (ovs *overlordSuite) TestNew(c *C) {
 	restore := patch.Mock(42, nil)
 	defer restore()
 
-	var setupStoreAuthContext auth.AuthContext
-	defer overlord.MockSetupStore(func(_ *state.State, ac auth.AuthContext) error {
-		setupStoreAuthContext = ac
-		return nil
-	})()
-
 	o, err := overlord.New()
 	c.Assert(err, IsNil)
 	c.Check(o, NotNil)
@@ -90,8 +84,9 @@ func (ovs *overlordSuite) TestNew(c *C) {
 	s.Get("patch-level", &patchLevel)
 	c.Check(patchLevel, Equals, 42)
 
-	// store was setup with an auth context
-	c.Check(setupStoreAuthContext, NotNil)
+	// store is setup
+	sto := snapstate.Store(s)
+	c.Check(sto, FitsTypeOf, &store.Store{})
 }
 
 func (ovs *overlordSuite) TestNewWithGoodState(c *C) {
@@ -156,15 +151,6 @@ func (ovs *overlordSuite) TestNewWithPatches(c *C) {
 	err = state.Get("patched", &b)
 	c.Assert(err, IsNil)
 	c.Check(b, Equals, true)
-}
-
-func (ovs *overlordSuite) TestNewWithSetupStoreError(c *C) {
-	defer overlord.MockSetupStore(func(*state.State, auth.AuthContext) error {
-		return errors.New("fake error")
-	})()
-
-	_, err := overlord.New()
-	c.Check(err, ErrorMatches, "fake error")
 }
 
 type witnessManager struct {
